automation_script/requirements.txt
anthropic
watchdog
python-docx
pypdf
openpyxl
python-dotenv
